fix: fixing and adding tests for ft_printf (beta)

+++ srcs/printf/ft_printf_d.c
@@ -30,15 +30,15 @@ int				main(int argc, char **argv)
 	else
 	{
 		if (argc == 4)
-			ret = ft_printf(argv[2], atol(argv[3]));
+			ret = ft_printf(argv[2], atoll(argv[3]));
 		else if (argc == 5)
-			ret = ft_printf(argv[2], atol(argv[3]), atol(argv[4]));
+			ret = ft_printf(argv[2], atoi(argv[3]), atoi(argv[4]));
 		else if (argc == 6)
-			ret = ft_printf(argv[2], atol(argv[3]), atol(argv[4]), atol(argv[5]));
+			ret = ft_printf(argv[2], atoi(argv[3]), atoi(argv[4]), atoi(argv[5]));
 		else if (argc == 7)
-			ret = ft_printf(argv[2], atol(argv[3]), atol(argv[4]), atol(argv[5]), atol(argv[6]));
+			ret = ft_printf(argv[2], atoi(argv[3]), atoi(argv[4]), atoi(argv[5]), atoi(argv[6]));
 		else if (argc == 8)
-			ret = ft_printf(argv[2], atol(argv[3]), atol(argv[4]), atol(argv[5]), atol(argv[6]), atol(argv[7]));
+			ret = ft_printf(argv[2], atoi(argv[3]), atoi(argv[4]), atoi(argv[5]), atoi(argv[6]), atoi(argv[7]));
 	}
 	printf("|%d", ret);
 	return (0);
